requirements.txt
openai
python-dotenv
juliacall
streamlit
plotly
